// src/utils/enum/pokemon-card-enums.ts
export enum PokemonType {
  COLORLESS = 'Colorless',
  DARKNESS = 'Darkness',
  DRAGON = 'Dragon',
  FAIRY = 'Fairy',
  FIGHTING = 'Fighting',
  FIRE = 'Fire',
  GRASS = 'Grass',
  LIGHTNING = 'Lightning',
  METAL = 'Metal',
  PSYCHIC = 'Psychic',
  WATER = 'Water',
}

export enum PokemonExpansion {
  BASE = 'Base',
  JUNGLE = 'Jungle',
  FOSSIL = 'Fossil',
  BASE_SET_2 = 'Base Set 2',
  TEAM_ROCKET = 'Team Rocket',
  GYM_HEROES = 'Gym Heroes',
  GYM_CHALLENGE = 'Gym Challenge',
  NEO_GENESIS = 'Neo Genesis',
  NEO_DISCOVERY = 'Neo Discovery',
}

export enum PokemonRarity {
  COMMON = 'COMMON',
  UNCOMMON = 'UNCOMMON',
  RARE = 'RARE',
}